+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 #. Deprecate `compute_settings` function.
+#. Ignore result of celery tasks as appropriate.
 
 1.2.6.1
 -------