+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ next
 #. Move to Django 1.6 support. Backwards incompatible.
 #. Use `django-layers-hr` to handle layering. The FOUNDRY['layers'] setting is now deprecated.
 #. Deprecate legacy handling for substring `_LAYER_` in photosize name.
+#. Add a `ViewProxy` model enabling views to appear in listings.
 
 1.2.6.1
 -------